Store and Cart functionality

+++ cart/views.py
@@ -2,7 +2,7 @@ from django.http import HttpResponse
 from django.shortcuts import render, redirect, get_object_or_404
 
 from cart.models import Cart, CartItem
-from products.models import Product
+from products.models import Product, Variation
 
 
 # Create your views here.
@@ -16,6 +16,18 @@ def _cart_id(request):  # this function retrieves or creates session keys used a
 
 def add_to_cart(request, product_id):
     product = Product.objects.get(id=product_id)  # get product
+    product_variation = []
+    if request.method == 'POST':
+        for item in request.POST:
+            key = item
+            value = request.POST[key]
+        try:
+            variation = Variation.objects.get(product=product, variation_category__iexact=key,
+                                              variation_value__iexact=value)
+            product_variation.append(variation)
+            print(product_variation)
+        except:
+            pass
 
     try:
         cart = Cart.objects.get(cart_id=_cart_id(request))  # get cart id through sessions
@@ -25,42 +37,67 @@ def add_to_cart(request, product_id):
         )
     cart.save()
 
-    try:
-        cart_item = CartItem.objects.get(product=product, cart=cart)
-        cart_item.quantity += 1
-        cart_item.save()
-    except CartItem.DoesNotExist:
+    is_cart_item_exist = CartItem.objects.filter(product=product, cart=cart).exists()
+    if is_cart_item_exist:
+        cart_item = CartItem.objects.filter(product=product, cart=cart)
+        # existing variations -> database
+
+        exist_var_list = []
+        id = []
+        for item in cart_item:
+            existing_variation = item.variations.all()
+            exist_var_list.append(list(existing_variation))
+            id.append(item.id)
+            print(exist_var_list)
+        # current variation -> product_variation
+
+        if product_variation in exist_var_list:
+            # Increase cart item quantity
+            index = exist_var_list.index(product_variation)
+            item_id = id[index]
+            item = CartItem.objects.get(product=product, id=item_id)
+            item.quantity += 1
+            item.save()
+
+        else:
+            item = CartItem.objects.create(product=product, quantity=1, cart=cart)
+            if len(product_variation) > 0:
+                item.variations.clear()
+                item.variations.add(*product_variation)
+            item.save()
+
+    else:
         cart_item = CartItem.objects.create(
             product=product,
             quantity=1,
             cart=cart
         )
+        if len(product_variation) > 0:
+            cart_item.variations.clear()
+            cart_item.variations.add(*product_variation)
         cart_item.save()
     return redirect('cart')
 
 
-def remove_from_cart(request, product_id):
+def remove_from_cart(request, product_id, cart_item_id):
     product = get_object_or_404(Product, id=product_id)  # get product
+    cart = Cart.objects.get(cart_id=_cart_id(request))
     try:
-        cart = Cart.objects.get(cart_id=_cart_id(request))  # get cart id through sessions
-    except Cart.DoesNotExist:
-        cart = Cart.objects.create(
-            cart_id=_cart_id(request)
-        )
-    cart.save()
-    try:
-        cart_item = CartItem.objects.get(product=product, cart=cart)
-        cart_item.quantity -= 1
-        cart_item.save()
-    except CartItem.DoesNotExist:
-        cart_item.delete()
+        cart_item = CartItem.objects.get(product=product, cart=cart, id=cart_item_id)
+        if cart_item.quantity > 1:
+            cart_item.quantity -= 1
+            cart_item.save()
+        else:
+            cart_item.delete()
+    except:
+        pass
     return redirect('cart')
 
 
-def remove_cart_item(request, product_id):
+def remove_cart_item(request, product_id, cart_item_id):
     product = get_object_or_404(Product, id=product_id)
     cart = Cart.objects.get(cart_id=_cart_id(request))
-    cart_item = CartItem.objects.get(product=product, cart=cart)
+    cart_item = CartItem.objects.get(product=product, cart=cart, id=cart_item_id)
     cart_item.delete()
     return redirect('cart')
